refactor: organize error classes to `errors.py`
